requirements.txt
openai==1.78.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
pgvector==0.4.1
python-dotenv==1.0.1
PyMuPDF==1.25.5
tiktoken==0.6.0
openai-agents==0.0.14
pydantic==2.6.1
ipython==8.22.2
tqdm==4.66.4
colorama==0.4.6
tenacity==8.3.0 